requirements.txt
streamlit
Pillow
pandas
gradio
openai
gpt_index
streamlit-chat 
streamlit_option_menu